pipeline/rag-orchestrator/requirements.txt
requests==2.31.0
jinja2==3.1.2
pytest==7.4.3
python-dotenv==1.0.0
anthropic>=0.18.0
flask>=3.0.0
flask-cors>=4.0.0
